requirements.txt
anthropic
langchain
langchain-anthropic
pydantic
pydantic-settings
python-dotenv
chromadb
web3
neynar-api-client
asyncpg==0.29.0
urllib3>=2.0.0
watchdog